Move CmdHelp to cli/help.go.
